backend/requirements/base.txt
Django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==24.2
drf-spectacular==0.27.2
psycopg2-binary==2.9.9
celery==5.3.6
redis==5.0.4
django-celery-beat==2.6.0
gspread==6.1.0
google-auth==2.29.0
firebase-admin==6.5.0
python-decouple==3.8
Pillow==10.3.0
